requirements.txt
pydantic
baml-py
python-dotenv
jupyterlab
openai